Format a data frame (tibble) as Airtable records

Gets a data frame and converts it into a list that can be passed to air_upload_records() to be uploaded to Airtable. This is required because the Airtable API does not accept tabular data and has certain restrictions about the format of the body request

Usage

air_tibble_to_records(
  this_tibble,
  id_fields = NULL,
  link_fields = NULL,
  max_records = 10
)

Arguments

this_tibble

A data frame with the fields to be uploaded to Airtable

id_fields

Character vector with the name of the column storing the id

link_fields

Character vector with the name of the columns that are link fields (fields that contain the id of another row's table)

max_records

Records are grouped into batches of max_records. Defaults to 10 which is the current limit imposed by the Airtable API.

Value

a list with records