Authenticate to a storage cloud provider

Usually used internally by other functions

cloud_storage_authenticate(provider, options)

Arguments

provider

cloud provider to use, either "gcs" or "aws"

options

named list with cloud provider options, see details

Details

Google Cloud Services

For Google Cloud Services ("gcs") options must be a list with the field service_account_key with the contents of the authentication json file you have downloaded from your Google Project.

This function uses googleCloudStorageR::gcs_auth under the hood to authenticate.

Examples

 # Google Cloud Services
if (FALSE) {
  authentication_details <- readLines("location_of_json_file.json")
  cloud_storage_authenticate(
    provider = "gcs",
    options = list(service_account_key = authentication_details,
                   bucket = "my-bucket"))
}