Download species body-length parameters

This function takes species, group of species or common names as arguments and returns a data frame containing weigth-length parameters from FishBase database.

get_fish_length(taxa, rank = NULL, country_code = NULL)

Arguments

taxa

The name of the species/groups from which retrieve lengths parameters.

rank

The taxonomical ranking, can be Species, Genus, Family, or comm_name. The latter is an abbreviation for the common name of a species/group.

country_code

The code of the country associated to FishBase database, default code include Timor-Leste and Indonesia. Full list at https://www.fishbase.se/country/ListOfCountryCodes.php.

Value

A data frame containing species length parameters

Examples

get_fish_length(
  taxa = "Sardine", rank = "comm_name",
  country_code = 626
)
#> Joining, by = c("Subfamily", "GenCode", "FamCode")
#> Joining, by = "FamCode"
#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")
#> Joining, by = c("Class", "ClassNum")
#> Joining, by = c("Subfamily", "GenCode", "FamCode")
#> Joining, by = "FamCode"
#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")
#> Joining, by = c("Class", "ClassNum")
#> Joining, by = "C_Code"
#> Joining, by = c("Subfamily", "GenCode", "FamCode")
#> Joining, by = "FamCode"
#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")
#> Joining, by = c("Class", "ClassNum")
#> Joining, by = c("Subfamily", "GenCode", "FamCode")
#> Joining, by = "FamCode"
#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")
#> Joining, by = c("Class", "ClassNum")
#> # A tibble: 24 × 42
#>    SpecCode Species AutoCtr StockCode FamCode PopLWRef LengthMin LengthMax Type 
#>       <int> <chr>     <int>     <int>   <int>    <int>     <dbl>     <dbl> <chr>
#>  1     1494 Herklo…    3769      1687      43    12145      NA        NA   FL   
#>  2     1494 Herklo…    3769      1687      43    12145      NA        NA   FL   
#>  3     1494 Herklo…    3770      1687      43    26587       5        10.5 FL   
#>  4     1494 Herklo…    3770      1687      43    26587       5        10.5 FL   
#>  5     1494 Herklo…    3771      1687      43    55787       5        14.5 FL   
#>  6     1494 Herklo…    3771      1687      43    55787       5        14.5 FL   
#>  7     1494 Herklo…   38463      1687      NA   111300       2.5      14.4 FL   
#>  8     1494 Herklo…   38463      1687      NA   111300       2.5      14.4 FL   
#>  9     1501 Amblyg…    3774      1694      43      823      NA        NA   NA   
#> 10     1501 Amblyg…    3774      1694      43      823      NA        NA   NA   
#> # … with 14 more rows, and 33 more variables: LmaxCompare <chr>, EsQ <chr>,
#> #   Number <int>, Sex <chr>, a <dbl>, aTL <dbl>, b <dbl>,
#> #   CoeffDetermination <dbl>, SEa <dbl>, SEb <dbl>, SDa <dbl>, SDb <dbl>,
#> #   LCLa <dbl>, UCLa <dbl>, LCLb <dbl>, UCLb <dbl>, Method <chr>,
#> #   SpecimenType <chr>, Locality <chr>, DataRef <int>, C_Code <chr>,
#> #   Comment <chr>, Entered <int>, DateEntered <dttm>, Modified <int>,
#> #   DateModified <dttm>, Expert <int>, DateChecked <dttm>, TS <dttm>, …