Download pelagic data system trips

Download pelagic data system trips

Usage

retrieve_pds_trips(
  prefix,
  secret = NULL,
  token = NULL,
  start_date = "2018-07-01",
  end_date = Sys.Date(),
  append_version = TRUE
)

Arguments

prefix

Name to be used as the prefix of the file names to be downloaded. Can be a path.

secret

Access secret code for the account

token

Access token for the account

start_date

Start date from which download trips information

end_date

Last date to download trips information

append_version

whether to append versioning information to the filename using add_version.

Value

A character vector with paths of the downloaded files

Examples

if (FALSE) { # \dontrun{
retrieve_pds_trips(
  prefix = "my-trips", secret = "abcXXXXXX",
  token = "123XXXXXX", start_date = "2018-07-01",
  end_date = Sys.Date()
)
# To download in a different path
dir.create("my-data-dir")
retrieve_pds_trips(
  prefix = "my-data-dir/my-trips", secret = "abcXXXXXX",
  token = "123XXXXXX", start_date = "2018-07-01",
  end_date = Sys.Date()
)
} # }