Download species body-length parameters

This function takes species, group of species or common names as arguments and returns a data frame containing weigth-length parameters from FishBase database.

get_fish_length(taxa, rank = NULL, country_code = NULL)

Arguments

taxa

The name of the species/groups from which retrieve lengths parameters.

rank

The taxonomical ranking, can be Species, Genus, Family, or comm_name. The latter is an abbreviation for the common name of a species/group.

country_code

The code of the country associated to FishBase database, default code include Timor-Leste and Indonesia. Full list at https://www.fishbase.se/country/ListOfCountryCodes.php.

Value

A data frame containing species length parameters

Examples

get_fish_length(
  taxa = "Sardine", rank = "comm_name",
  country_code = 626
)
#> Joining with `by = join_by(Subfamily, GenCode, FamCode)`
#> Joining with `by = join_by(FamCode)`
#> Joining with `by = join_by(Order, Ordnum, Class, ClassNum)`
#> Joining with `by = join_by(Class, ClassNum)`
#> Joining with `by = join_by(SpecCode)`
#> Joining with `by = join_by(C_Code)`
#> Joining with `by = join_by(SpecCode)`
#> Joining with `by = join_by(SpecCode)`
#> # A tibble: 24 × 42
#>    Species SpecCode AutoCtr StockCode FamCode PopLWRef LengthMin LengthMax Type 
#>    <chr>      <int>   <int>     <int>   <int>    <int>     <dbl>     <dbl> <chr>
#>  1 Amblyg…     1501    3774      1694      43      823      NA        NA   NA   
#>  2 Amblyg…     1501    3774      1694      43      823      NA        NA   NA   
#>  3 Amblyg…     1501    3775      1694      43     1439       8        22.9 TL   
#>  4 Amblyg…     1501    3775      1694      43     1439       8        22.9 TL   
#>  5 Amblyg…     1501    3776      1694      43     1439       9.1      21.8 TL   
#>  6 Amblyg…     1501    3776      1694      43     1439       9.1      21.8 TL   
#>  7 Amblyg…     1501    3777      1694      43     1439       4.2      22.9 TL   
#>  8 Amblyg…     1501    3777      1694      43     1439       4.2      22.9 TL   
#>  9 Amblyg…     1501    3778      1694      43     1488      NA        NA   NA   
#> 10 Amblyg…     1501    3778      1694      43     1488      NA        NA   NA   
#> # ℹ 14 more rows
#> # ℹ 33 more variables: LmaxCompare <chr>, EsQ <chr>, Number <int>, Sex <chr>,
#> #   a <dbl>, aTL <dbl>, b <dbl>, CoeffDetermination <dbl>, SEa <dbl>,
#> #   SEb <dbl>, SDa <dbl>, SDb <dbl>, LCLa <dbl>, UCLa <dbl>, LCLb <dbl>,
#> #   UCLb <dbl>, Method <chr>, SpecimenType <chr>, Locality <chr>,
#> #   DataRef <int>, C_Code <chr>, Comment <chr>, Entered <int>,
#> #   DateEntered <dttm>, Modified <int>, DateModified <dttm>, Expert <int>, …