Download species body-length parameters

This function takes species, group of species or common names as arguments and returns a data frame containing weigth-length parameters from FishBase database.

get_fish_length(taxa, rank = NULL, country_code = NULL)

Arguments

taxa
The name of the species/groups from which retrieve lengths parameters.

rank
The taxonomical ranking, can be Species, Genus, Family, or comm_name. The latter is an abbreviation for the common name of a species/group.

country_code
The code of the country associated to FishBase database, default code include Timor-Leste and Indonesia. Full list at https://www.fishbase.se/country/ListOfCountryCodes.php.

Value

A data frame containing species length parameters

Examples

get_fish_length(
  taxa = "Sardine", rank = "comm_name",
  country_code = 626
)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/comnames_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 324211 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (34): autoctr, ComName, Transliteration, StockCode, SpecCode, C_Code, La...
#> dbl  (1): ComNamesRefNo
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 27.13703 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/species_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 35331 Columns: 101
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (63): SpecCode, Genus, Species, Author, FBname, PicPreferredName, PicPre...
#> dbl (36): SpeciesRefNo, FamCode, GenCode, SubGenCode, Fresh, Brack, Saltwate...
#> lgl  (2): AmphibiousRef, Complete
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 8.165497 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/genera_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 11863 Columns: 64
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (35): GenCode, GenName, GenAuthorYear, GenAuthor, GenYear, GenInEx, Gen...
#> dbl  (24): GenRefno, FamCode, DepthRangeShallow, FB_SppNb, CofF_SppNb, FoW_S...
#> lgl   (2): Brackish, Freshwater
#> dttm  (3): Dateentered, Datemodified, Datechecked
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 1.905385 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/families_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 1288 Columns: 81
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (50): FamCode, SortID, Family, AuthorYear, Qualification, CommonName, F...
#> dbl  (23): Species, SexualDimorphism, FossilRef, ValidGenera, Complete, Aspi...
#> lgl   (5): CAS_LIN, ZooBank_ID, CommonName_Chinese, CommonName_Chinese_u, Co...
#> dttm  (3): DateEntered, DateModified, DateChecked
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 0.360585 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/orders_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 113 Columns: 56
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (35): Ordnum, Order, Synonym, CommonName, BodyShapeI, Brackish, Freshwa...
#> dbl  (12): OrderRefNo, SortNo, AuthorYear, Marine, FamCount, ClassNum, ComAn...
#> lgl   (7): CAS_LIN, ZooBank_ID, DateChecked, CommonName_Spanish, CommonName_...
#> dttm  (2): DateEntered, DateModified
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 0.2800281 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/classes_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 14 Columns: 16
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (7): Class, CommonName, BodyShapeI, WaterSalinity, SuperClass, Subclass,...
#> dbl (5): ClassNum, ClassRefNo, SortNo, Synonym, SpeciesCount
#> lgl (4): AuthorYear, CAS_LIN, ZooBank_ID, TS
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 0.08759713 secs)
#> Joining, by = c("Subfamily", "GenCode", "FamCode")
#> Joining, by = "FamCode"
#> Joining, by = c("Order", "Ordnum", "Class", "ClassNum")
#> Joining, by = c("Class", "ClassNum")
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/country_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 196203 Columns: 36
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (24): autoctr, Stockcode, C_Code, SpecCode, CountryRefNo, Status, Curren...
#> dbl (12): AlsoRef, Freshwater, Saltwater, Land, RefImportance, ExVesselPrice...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 14.74628 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/countref_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 2344 Columns: 204
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr   (85): PAESE, Note, C_Code, ABB, ISO, AdminCountry, ISO2Alpha, ISO3Alph...
#> dbl  (113): ISO1Num, SAUP_Code, Used, LatDeg, LatMin, LongDeg, LongMin, Pola...
#> lgl    (5): RefFresh2, RefFresh3, NatFishDB2, TradeNames, DateChecked
#> dttm   (1): LastUpdate
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 1.133184 secs)
#> Joining, by = "C_Code"
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/poplw_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 19692 Columns: 37
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr (22): AutoCtr, SpecCode, FamCode, PopLWRef, LengthMin, Type, LmaxCompare...
#> dbl (15): StockCode, LengthMax, Number, a, aTL, b, CoeffDetermination, SEa, ...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 1.514299 secs)
#> Importing /Users/runner/Library/Application Support/org.R-project.R/R/rfishbase/popll_fb_2104.tsv.bz2 in 1000000 line chunks:
#> Rows: 26873 Columns: 22
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr   (7): autoctr, Length1, Length2, a, Type, Sex, Comment
#> dbl  (11): StockCode, SpecCode, b, LengthMin, LengthMax, Number, r2, DataRef...
#> lgl   (1): TS
#> dttm  (3): DateEntered, DateModified, DateChecked
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> 	...Done! (in 1.330945 secs)
#> # A tibble: 24 × 42
#>    SpecCode Species AutoCtr StockCode FamCode PopLWRef LengthMin LengthMax Type 
#>    <chr>    <chr>   <chr>       <dbl> <chr>   <chr>    <chr>         <dbl> <chr>
#>  1 1494     Herklo… 3769         1687 43      12145    NA             NA   FL   
#>  2 1494     Herklo… 3769         1687 43      12145    NA             NA   FL   
#>  3 1494     Herklo… 3770         1687 43      26587    5              10.5 FL   
#>  4 1494     Herklo… 3770         1687 43      26587    5              10.5 FL   
#>  5 1494     Herklo… 3771         1687 43      55787    5              14.5 FL   
#>  6 1494     Herklo… 3771         1687 43      55787    5              14.5 FL   
#>  7 1494     Herklo… 38463        1687 NA      111300   2.5            14.4 FL   
#>  8 1494     Herklo… 38463        1687 NA      111300   2.5            14.4 FL   
#>  9 1501     Amblyg… 11380        1694 NA      93920    12             21   FL   
#> 10 1501     Amblyg… 11380        1694 NA      93920    12             21   FL   
#> # … with 14 more rows, and 33 more variables: LmaxCompare <chr>, EsQ <chr>,
#> #   Number <dbl>, Sex <chr>, a <dbl>, aTL <dbl>, b <dbl>,
#> #   CoeffDetermination <dbl>, SEa <dbl>, SEb <dbl>, SDa <dbl>, SDb <dbl>,
#> #   LCLa <dbl>, UCLa <dbl>, LCLb <dbl>, UCLb <dbl>, Method <chr>,
#> #   SpecimenType <chr>, Locality <chr>, DataRef <chr>, C_Code <chr>,
#> #   Comment <chr>, Entered <chr>, DateEntered <chr>, Modified <chr>,
#> #   DateModified <chr>, Expert <chr>, DateChecked <chr>, TS <chr>, …