Join length-weights and nutritional parameters info to preprocessed landings

The function integrate nutritional info from the get_nutrients_table function and length-weight info from FIshBase databse to preprocessed surveys data and convert catch labels according to the FAO nomenclature (http://www.fao.org/fishery/statistics/global-production/3/en).

Usage

join_weights(data, metadata, rfish_tab, nutrients_table)

Arguments

data

The survey landings data frame

metadata

Metadata tables

rfish_tab

Table with length weight parameters

nutrients_table

Table with nutritional parameters

Value

A new landings data frame including length-weights info

Details

The length types used to calculate weight in fish catches include total length (TL) in survey version 1 and fork length (FL) in survey version 2 with the exception of the group SRX (Myliobatiformes), which uses disk width (WD). Weights of Non-fish groups are calculated according to carapace width (CW) in crabs, mantel length (ML) in Cephalopoda and shell length (ShL) in Bivalvia, carapace length (CL) in lobsters.