Validate surveys' catch parameters

This function takes a preprocessed landings' matrix and uses univariate techniques (see univOutl::LocScaleB) for the identification of outliers in the distribution of the number of individuals per catch and their size. The function returns a data frame with the survey id, the alert number and a nested column species_group containing validated catches parameters.

validate_catch_params(data, method = NULL, k_ind = NULL, k_length = NULL)

Arguments

data
A preprocessed data frame

method
character identifying how to estimate the scale of the distribution. Available choices are:

method='IQR' for using the Inter-Quartile Range, i.e. Q3-Q1;

method='IDR' for using the Inter-Decile Range; i.e. P90-P10

method='MAD' for using the Median Absolute Deviation;

method='Gini' robust scale estimate based on Gini's Mean Difference (see GiniMd);

method='ScaleTau2' robust tau-estimate of univariate scale, as proposed by Maronna and Zamar (2002) (see alsoscaleTau2);

method='Qn' for using the Qn estimator proposed by Rousseeuw and Croux (1993) (see also Qn);

method='Sn' for using the Sn estimator proposed by Rousseeuw and Croux (1993) (see also Sn).

When method='dQ' the estimated scale for the left tail is (Q2-Q1)/0.6745, while for the right tail it is considered (Q3-Q2)/0.6745 (Q2 is the median); this double estimate should be able to account for slight skewness.

When method='dD' the estimated scale for the left tail is (P50-P10)/1.2816, while for the right tail it is considered (P90-P50)/1.2816 (P50 is the median); this double estimate should be able to account for skewness.

Finally, when method='AdjOut', bounds are based on the adjusted outlyingness method as proposed by Hubert and Van der Veeken (2008).

k_ind
Extension of bounds for the number of individuals

k_length
Extension of bounds for the catch length

Value

A data frame containing the validated catches parameters.

Examples

if (FALSE) {
  pars <- read_config()
  landings <- get_merged_landings(pars)
  validate_catch_params(landings,method="MAD",k=13)
}