Validate surveys' total catch values

This function takes a preprocessed landings' matrix and uses univariate techniques (see univOutl::LocScaleB) for the identification of outliers in the distribution of the total catch values associated to surveys.

validate_catch_price(data, method = NULL, k = NULL)

Arguments

data

A preprocessed data frame

method

character identifying how to estimate the scale of the distribution. Available choices are:

method='IQR' for using the Inter-Quartile Range, i.e. Q3-Q1;

method='IDR' for using the Inter-Decile Range; i.e. P90-P10

method='MAD' for using the Median Absolute Deviation;

method='Gini' robust scale estimate based on Gini's Mean Difference (see GiniMd);

method='ScaleTau2' robust tau-estimate of univariate scale, as proposed by Maronna and Zamar (2002) (see alsoscaleTau2);

method='Qn' for using the Qn estimator proposed by Rousseeuw and Croux (1993) (see also Qn);

method='Sn' for using the Sn estimator proposed by Rousseeuw and Croux (1993) (see also Sn).

When method='dQ' the estimated scale for the left tail is (Q2-Q1)/0.6745, while for the right tail it is considered (Q3-Q2)/0.6745 (Q2 is the median); this double estimate should be able to account for slight skewness.

When method='dD' the estimated scale for the left tail is (P50-P10)/1.2816, while for the right tail it is considered (P90-P50)/1.2816 (P50 is the median); this double estimate should be able to account for skewness.

Finally, when method='AdjOut', bounds are based on the adjusted outlyingness method as proposed by Hubert and Van der Veeken (2008).

k

Nonnegative constant that determines the extension of bounds. Commonly used values are 2, 2.5 and 3 (default).

Value

A data frame containing validated catch values.

Examples

if (FALSE) {
pars <- read_config()
landings <- get_merged_landings(pars)
validate_catch_value(landings, method = "MAD", k = 13)
}