Validate landings

Downloads the preprocessed version of the data from cloud storage services and validates a range of information so that it can be safely used for analysis. By default the function uses the method of the median absolute deviation (MAD) for outliers identification.

validate_landings(log_threshold = logger::DEBUG)

Arguments

log_threshold

The (standard Apache logj4) log level used as a threshold for the logging infrastructure. See logger::log_levels for more details

Value

no outputs. This function is used for it's side effects

Details

The parameters needed in the config file are those required for preprocess_landings_step_1() or preprocess_landings_step_2(), preprocess_metadata_tables() combined, as well as parameters needed to outliers identification that are hrs, methodandk`.