Validate surveys' temporal parameters

This function takes a preprocessed landings' matrix and validate temporal info associated to each survey.

validate_surveys_time(data, hrs = NULL, submission_delay)

Arguments

data

A preprocessed data frame

hrs

Limit of trip duration in hours to be considered a valid catch session.

submission_delay

Limit for maximum difference (in days) between the survey submission date and the recorded landing datw

Value

A list containing data frames with validated catch dates and catch duration.

Examples

if (FALSE) {
  pars <- read_config()
  landings <- get_merged_landings(pars)
  validate_surveys_time(landings,hrs =18)
}